refactor(mvm-build): migrate finalize_flake_job + finalize_install_job into BuilderVmRuntime
Plan 97 Phase C, third migration after stage_job_dir (#434) and
read_job_result (#435). Lifts the hypervisor-agnostic artifact
finalisation out of libkrun_builder.rs so the future VzBuilderVm path
reuses it verbatim.

Migrated:
- INSTALL_RESULT_FILENAME const
- read_last_bytes_of (tail of nix-stderr.log capture)
- finalize_flake_job (rootfs + kernel + revision-hash extraction)
- read_revision_hash + extract_nix_store_hash (store-path parser)
- finalize_install_job (install report + sealed-volume validation)
- InstallResultReport struct
- 13 associated unit tests

All items are pure host-side I/O against the /job and /out virtio-fs
shares, which libkrun and Vz attach identically. The two production
call sites in libkrun_builder.rs (BuilderJob match dispatch) keep
their existing names via use-import.

serde::Deserialize import drops from libkrun_builder.rs (both JobResult
and InstallResultReport now live elsewhere).
